Gère le lancement du script en PR
