--- .github/actions/reopen-issue-with-comment/action.yaml
@@ -1,13 +1,0 @@
-name: "Reopen and comment"
-description: "Reopen an issue if closed and add a comment"
-inputs:
-  token:
-    description: "GITHUB_TOKEN or a repo scoped PAT."
-    default: ${{ github.token }}
-  issue number:
-    description: "The number of the issue or pull request in which to create a comment."
-  comment:
-    description: "The comment body."
-runs:
-  using: "node16"
-  main: "./main.js"
