--- .github/actions/reopen-issue-with-comment/readme.md
@@ -1,7 +1,0 @@
-This action can be tested locally with the following command line:
-
-```bash
-INPUT_TOKEN=TOKEN INPUT_ISSUE_NUMBER=2945 INPUT_COMMENT='Hi<br />World!' GITHUB_REPOSITORY=betagouv/aides-jeunes node main.js
-```
-
-`TOKEN` should be replaced by a real token generated from https://github.com/settings/tokens/ with full control on repository.
