+++ package.json
@@ -30,7 +30,7 @@
     "start": "cd dist-server && NODE_ENV=production node ./backend/server.js",
     "stats": "ts-node ./backend/lib/stats",
     "test": "NODE_OPTIONS=--experimental-vm-modules NODE_PATH=. jest --testTimeout=20000",
-    "tools:check-links-validity": "ts-node ./tools/check-links-validity.ts --ci",
+    "tools:check-links-validity": "ts-node ./tools/check-links-validity.ts",
     "tools:cleaner": "ts-node ./tools/cleaner.ts",
     "tools:evaluate-benefits": "ts-node ./tools/evaluate-benefits",
     "tools:generate-missing-institutions-aides-velo": "ts-node ./tools/generate-missing-institutions-aides-velo",
